Refactor event calendar flow with new modular components and remove repeated date conversion function
