+++ src/screens/Events/EventDetails.js
@@ -1,12 +1,11 @@
 import React, { Component } from 'react';
 import { connect } from 'react-redux';
-import _ from 'lodash';
 import { Alert, Button, NavBar, FilterList, ButtonLayout, Icon } from '@/components';
 import QRCode from 'react-native-qrcode-svg';
 import QRCodeScanner from 'react-native-qrcode-scanner';
 import { MemberPanel } from '@/utils/render';
-import { months } from '@/data/DateItems';
 import { EventForm } from '@/data/FormData';
+import { convertNumToDate } from '@/utils/events';
 import { deleteEvent, editEvent, checkIn, rsvp, pageLoad, getAllMemberAccounts } from '@/ducks';
 import {
 	View,
@@ -40,13 +39,6 @@ class EventDetails extends Component {
 
 		if (activeUser.privilege && activeUser.privilege.board)
 			getAllMemberAccounts();
-	}
-
-	convertNumToDate(date) {
-		if (!date) return '';
-		let tempDate = date.split('-');
-
-		return `${months[Number(tempDate[1]) - 1]} ${tempDate[2]}`;
 	}
 
 	onSuccess = (e) => {
@@ -438,7 +430,7 @@ class EventDetails extends Component {
 								name = 'md-calendar'
 								size = { iconSize }
 								color = '#000' />
-							<Text style = { [text, textColor] }>{ this.convertNumToDate(date) }</Text>
+							<Text style = { [text, textColor] }>{ convertNumToDate(date) }</Text>
 						</View>
 						<View style = { iconContainer }>
 							<Icon
