+++ src/screens/General/Dashboard/index.js
@@ -1,4 +1,3 @@
-export { default as EventsList } from './EventsList';
 export { default as FavoriteCommittees } from './FavoriteCommittees';
 export { default as Leaderboard } from './Leaderboard';
 export { default as Dashboard } from './Dashboard';
